perf(config-panel): virtualize the device list on HA 2026.7+
Renders the device list through ha-list-virtualized above 100 devices, so only visible rows hit the DOM.

Feature-detected via customElements.get: the element does not exist before HA 2026.7, and an unknown element would leave the list silently blank rather than error. Below the threshold and on older HA the plain grouped list is used, so the minimum stays HA 2026.3.

The row markup moves into its own hm-device-row element: the virtualizer renders rows inside its own shadow root, where the list's styles do not reach. Both render paths share it.
